chore(project): rename ui folder to container
