+++ ui/src/containers/MarkdownImage.stories.tsx
