+++ ui/src/containers/MarkdownImage.tsx
