+++ ui/src/containers/ProjectHighlightCarousel.stories.tsx
