+++ ui/src/containers/ProjectHighlightCarousel.tsx
