+++ ui/src/containers/SimpIcon.stories.tsx
