+++ ui/src/containers/Social.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import { SocialLinks } from '@/ui/SocialLinks';
+import { SocialLinks } from '@/containers/SocialLinks';
 import { SOCIAL_LINKS } from '@/lib/consts';
 import { withTheme } from '@/lib/storybook-utils';
