+++ ui/src/containers/customIcons/LogoIcon.tsx
